Ensure u64 to usize cast is lossless
This commit casts the `usize` values to `u64`, performs the modulus
operation, and then converts the small remainder back to `usize`. This
calculation is more likely to be lossless compared to the previous
method, which casts a potentially large value directly to `usize` before
performing the modulus which might truncate the value on 32-bit systems.

+++ src/ser.rs
@@ -94,7 +94,7 @@ where
     writer.write_all(&len.to_le_bytes())?;
     io::copy(reader, writer)?;
 
-    let remainder = len as usize % PAD_LEN;
+    let remainder = (len % PAD_LEN as u64) as usize;
     if remainder > 0 {
         let buf = [0u8; PAD_LEN];
         let padding = PAD_LEN - remainder;
